requirements.txt
pandas==1.2.4
SoundFile==0.10.3.post1
torchinfo==0.1.4
tqdm==4.61.0